validation/src/main/java/com/spotify/protoman/validation/rules/FieldLabelRule.java
/*-
 * -\-\-
 * protoman-validation
 * --
 * Copyright (C) 2016 - 2018 Spotify AB
 * --
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * 
 *      http://www.apache.org/licenses/LICENSE-2.0
 * 
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * -/-/-
 */

package com.spotify.protoman.validation.rules;

import com.spotify.protoman.descriptor.FieldDescriptor;
import com.spotify.protoman.validation.ComparingValidationRule;
import com.spotify.protoman.validation.ValidationContext;
import com.spotify.protoman.validation.ViolationType;

/**
 * Verifies that field labels are not changed in an incompatible way.
 */
public class FieldLabelRule implements ComparingValidationRule {

  private FieldLabelRule() {
  }

  public static FieldLabelRule create() {
    return new FieldLabelRule();
  }

  @Override
  public void fieldChanged(final ValidationContext ctx,
                           final FieldDescriptor current,
                           final FieldDescriptor candidate) {
    // If label is "required" changing it is wire-incompatible.
    if (current.isRequired() != candidate.isRequired()) {
      ctx.report(
          ViolationType.WIRE_INCOMPATIBILITY_VIOLATION,
          "field label changed to/from required"
      );
    } else if (current.isRepeated() != candidate.isRepeated()) {
      // Changing between repeated and optional is wire-compatible but will affect generated
      // source code.
      ctx.report(
          ViolationType.GENERATED_SOURCE_CODE_INCOMPATIBILITY_VIOLATION,
          "field label changed"
      );
    }
  }
}
